% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consort.R
\name{consort}
\alias{consort}
\title{\link[consort]{consort_plot}}
\description{
\link[consort]{consort_plot}
}
\examples{
# example from \pkg{consort} vignette
library(consort)
data(dispos.data)
cs = consort_plot(data = dispos.data, orders = c(
  trialno = 'Population',
  exclusion = 'Excluded',
  trialno = 'Allocated',
  subjid_notdosed = 'Not dosed',
  followup = 'Followup'
), side_box = c('exclusion', 'subjid_notdosed'), cex = 0.9)
list(
 '`consort`' = cs, 
 '`grob`' = build_grid(cs)
) |> fastmd::render2html()
  
}
