% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consort.R
\name{consort_rx}
\alias{consort_rx}
\title{Finetune Function \link[consort]{consort_plot}}
\usage{
consort_rx(data, orders, sidebox_pattern = "^sidebox_", ...)
}
\arguments{
\item{data}{\link[base]{data.frame}, see function \link[consort]{consort_plot}}

\item{orders}{\link[base]{character} \link[base]{vector}, see function \link[consort]{consort_plot}}

\item{sidebox_pattern}{\link[base]{regex}}

\item{...}{additional parameters of \link[consort]{consort_plot}}
}
\value{
Function \code{\link[=consort_rx]{consort_rx()}} returns a
\link[consort]{consort_plot} object.
}
\description{
Some minor touch-ups of function \link[consort]{consort_plot}.
}
\details{
Function \code{\link[=consort_rx]{consort_rx()}}
\itemize{
\item{finds the argument \code{side_box} of function \link[consort]{consort_plot}, from the argument of \code{orders}.}
}
}
\keyword{internal}
