% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0PACKAGE.R
\docType{package}
\name{grid.tzh-package}
\alias{grid.tzh}
\alias{grid.tzh-package}
\title{grid.tzh: Tingting's Extension to grid Package}
\description{
T. Zhan's extension to package 'grid', as well as applications.
}
\author{
\strong{Maintainer}: Tingting Zhan \email{tingtingzhan@gmail.com} (\href{https://orcid.org/0000-0001-9971-4844}{ORCID})

}
\keyword{internal}
