% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unit.R
\name{grid_unit}
\alias{grid_unit}
\alias{units<-.unit}
\alias{units.unit}
\alias{>.unit}
\alias{>=.unit}
\alias{<.unit}
\alias{<=.unit}
\alias{!=.unit}
\alias{==.unit}
\title{\link[grid]{unit}}
\usage{
\method{units}{unit}(x) <- value

\method{units}{unit}(x)

\method{>}{unit}(e1, e2)

\method{>=}{unit}(e1, e2)

\method{<}{unit}(e1, e2)

\method{<=}{unit}(e1, e2)

\method{!=}{unit}(e1, e2)

\method{==}{unit}(e1, e2)
}
\arguments{
\item{x, e1, e2}{\link[grid]{unit}}

\item{value}{..}
}
\description{
\link[grid]{unit}
}
\note{
Accepted units for function \link[grid]{unit} can be found in
\verb{R source/src/library/grid/src/unit.c} file.

Only \strong{length} units for plot are accepted.
Weight units (such as kg or lb) are not accepted!
}
\examples{
library(grid)
(x = unit(10, units = 'cm'))
units(x) = 'inches'
x

(x1 = unit(rnorm(10), units = 'inches'))
(x2 = unit(rnorm(10), units = 'cm'))
x1 > x2


}
