% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsub_text_label.R
\name{gsub_text_label}
\alias{gsub_text_label}
\alias{gsub_text_label.default}
\alias{gsub_text_label.grob}
\alias{gsub_text_label.gList}
\title{Modify \verb{$label} of a \code{'text'} \link[grid]{grob}}
\usage{
gsub_text_label(pattern, replacement, x, ...)

\method{gsub_text_label}{default}(pattern, replacement, x, ...)

\method{gsub_text_label}{grob}(pattern, replacement, x, ...)

\method{gsub_text_label}{gList}(pattern, replacement, x, ...)
}
\arguments{
\item{pattern, replacement, ...}{see function \link[base]{gsub}}

\item{x}{see \strong{Usage}}
}
\description{
Modify \verb{$label} of a \code{'text'} \link[grid]{grob}
}
\note{
tzh may move function \code{\link[=gsub_text_label]{gsub_text_label()}} to a more basic package, maybe named \pkg{grid.tzh}, in future.
}
\keyword{internal}
