% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consort.R
\name{md_.consort}
\alias{md_.consort}
\title{Markdown Lines for \link[consort]{consort_plot}}
\usage{
\method{md_}{consort}(x, ...)
}
\arguments{
\item{x}{a \link[consort]{consort_plot}}

\item{...}{additional parameters of function \link[fastmd]{md_.default}}
}
\description{
Markdown Lines for \link[consort]{consort_plot}
}
\examples{
# example from \pkg{consort} vignette
library(consort)
data(dispos.data)
cs1 = consort_plot(data = dispos.data, orders = c(
  trialno = 'Population',
  exclusion = 'Excluded',
  trialno = 'Allocated',
  subjid_notdosed = 'Not dosed',
  followup = 'Followup'
), side_box = c('exclusion', 'subjid_notdosed'), cex = 0.9)
cs2 = cs1 |> 
  build_grid()
list(consort1 = cs1, consort2 = cs2) |> 
  fastmd::render2html()
}
\keyword{internal}
