% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_grid.R
\name{md_.gDesc}
\alias{md_.gDesc}
\title{Markdown Lines for \code{gDesc}}
\usage{
\method{md_}{gDesc}(x, xnm, ...)
}
\arguments{
\item{x}{\code{gDesc}}

\item{xnm}{..}

\item{...}{..}
}
\description{
Markdown Lines for \code{gDesc}
}
\examples{
# example from \pkg{consort} vignette
library(consort)
data(dispos.data)
out = consort_plot(data = dispos.data, orders = c(
  trialno = 'Population',
  exclusion = 'Excluded',
  trialno = 'Allocated',
  subjid_notdosed = 'Not dosed',
  followup = 'Followup'
), side_box = c('exclusion', 'subjid_notdosed'), cex = 0.9)
p = out |> build_grid()
attr(p, which = 'text') <- Sprintf_consort()
library(rmd.tzh); list(
  '`consort`' = p
) |> render_(file = 'gDesc')
}
\keyword{internal}
