% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/md_grid.R
\name{md_.gList}
\alias{md_.gList}
\title{Markdown Lines for \code{gList}}
\usage{
\method{md_}{gList}(x, xnm, ...)
}
\arguments{
\item{x}{\code{gList}}

\item{xnm}{..}

\item{...}{..}
}
\description{
Markdown Lines for \code{gList}
}
\examples{
# see ?venn
}
\keyword{internal}
