% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn.R
\name{md_.VennDiagram}
\alias{md_.VennDiagram}
\title{Markdown Lines for \code{VennDiagram}}
\usage{
\method{md_}{VennDiagram}(x, ...)
}
\arguments{
\item{x}{a \code{VennDiagram}}

\item{...}{additional parameters of function \link[rmd.tzh]{md_.default}}
}
\description{
Markdown Lines for \code{VennDiagram}
}
\examples{
# see ?venn
}
\keyword{internal}
