% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consort.R
\name{paste_nna_}
\alias{paste_nna_}
\title{Element-Wise Concatenation of Non-Missing Entries}
\usage{
paste_nna_(..., sep = "; ")
}
\arguments{
\item{...}{two or more \link[base]{character} \link[base]{vector}s}

\item{sep}{\link[base]{character} scalar}
}
\value{
Function \link{paste_nna_} returns a \link[base]{character} \link[base]{vector}.
}
\description{
Element-wise concatenation of non-\link[base]{missing} entries.
}
\details{
Function \link{paste_nna_} can be used to combine two or more reasons in a \CRANpkg{consort} diagram.
}
\examples{
x = c(NA_character_, 'x1', 'x2')
y = c(NA_character_, 'y1', NA_character_)
paste(x, y, sep = '; ')
paste_nna_(x, y, sep = '; ')
}
\keyword{internal}
