% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.gList.R
\name{plot.gList}
\alias{plot.gList}
\title{Plot \link[grid]{gList} Object}
\usage{
\method{plot}{gList}(x, ...)
}
\arguments{
\item{x}{a \link[grid]{gList}}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \code{\link[=plot.gList]{plot.gList()}} does not have a returned value.
}
\description{
Plot \link[grid]{gList} Object
}
\note{
We don't have \code{grid:::plot.gList}, as of R version 4.4.3.
}
