% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn.R
\name{print.venn}
\alias{print.venn}
\title{\link{print.venn}}
\usage{
\method{print}{venn}(x, ...)
}
\arguments{
\item{x}{..}

\item{...}{..}
}
\description{
\link{print.venn}
}
\note{
So that I can simply use \link[rmd.tzh]{md_.default} :)
}
