% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn.R
\name{set_xor}
\alias{set_xor}
\title{Exclusive-OR Elements}
\usage{
set_xor(...)
}
\arguments{
\item{...}{two named R objects of the same \link[base]{typeof}}
}
\value{
The function \code{\link[=set_xor]{set_xor()}} returns a \link[stats]{listof} of \link[base]{vector}s.
}
\description{
Exclusive-OR elements in two vectors.
}
\details{
The function \code{\link[=set_xor]{set_xor()}} returns the exclusive-OR elements in each of the sets, which is slow
and only intended for end-user.
}
\examples{
set_xor(a = 1:5, b = 3:7)
set_xor(a = 1:5, b = 1:3)
}
\seealso{
\link[base]{setequal}
}
\keyword{internal}
