% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consort.R
\name{sidebox}
\alias{sidebox}
\title{sidebox}
\usage{
sidebox(pattern = "^sidebox_", data, ...)
}
\arguments{
\item{pattern}{\link[base]{regex} of \emph{ineligibility} criterion}

\item{data}{\link[base]{data.frame}}

\item{...}{additional parameters, currently not in use}
}
\value{
Function \code{\link[=sidebox]{sidebox()}} returns a \link[base]{matrix}
}
\description{
sidebox
}
\details{
Function \code{\link[=sidebox]{sidebox()}} converts ineligibility criterion in \emph{side box} into
a \link[base]{logical} \link[base]{matrix} of \strong{eligibility} criterion.
}
\keyword{internal}
