% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consort.R
\name{consort_plot_rx}
\alias{consort_plot_rx}
\title{Finetune Function \link[consort]{consort_plot}}
\usage{
consort_plot_rx(
  data,
  orders,
  sidebox_pattern = "^sidebox_",
  ...,
  top = NULL,
  bottom = NULL,
  left = NULL,
  right = NULL
)
}
\arguments{
\item{data}{\link[base]{data.frame}, see function \link[consort]{consort_plot}}

\item{orders}{\link[base]{character} \link[base]{vector}, see function \link[consort]{consort_plot}}

\item{sidebox_pattern}{\link[base]{regex}}

\item{...}{additional parameters of \link[consort]{consort_plot}}

\item{top, bottom, left, right}{see function \link[gridExtra]{arrangeGrob}}
}
\value{
Function \code{\link[=consort_plot_rx]{consort_plot_rx()}} returns a
\link[grid]{grob} object.
}
\description{
Some minor touch-ups of function \link[consort]{consort_plot}.
}
\details{
Function \code{\link[=consort_plot_rx]{consort_plot_rx()}}
\itemize{
\item{finds the argument \code{side_box} of function \link[consort]{consort_plot}, from the argument of \code{orders}.}
\item{adds top, bottom, left and/or right labels, via function \link[gridExtra]{arrangeGrob}.}
\item{returns a \link[grid]{grob} object.}
}
}
