% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn.R
\name{md_.venn}
\alias{md_.venn}
\title{Markdown Lines for \code{venn}}
\usage{
\method{md_}{venn}(x, ...)
}
\arguments{
\item{x}{\code{venn}}

\item{...}{additional parameters of function \link[rmd.tzh]{md_.default}}
}
\description{
Markdown Lines for \code{venn}
}
\examples{
# see ?venn
}
\keyword{internal}
