% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/venn.R
\name{md_.VennDiagram}
\alias{md_.VennDiagram}
\title{Fast Mark Down Lines for \code{VennDiagram}}
\usage{
\method{md_}{VennDiagram}(x, xnm, ...)
}
\arguments{
\item{x}{a \code{VennDiagram}}

\item{xnm}{..}

\item{...}{..}
}
\description{
Fast Mark Down Lines for \code{VennDiagram}
}
\keyword{internal}
